feat: remove resize
